split up state module
