+++ gerritbot/src/state.rs
@@ -1,10 +1,6 @@
-use std::borrow::Borrow;
 use std::collections::HashMap;
-use std::collections::HashSet;
-use std::fmt::{Display, Formatter, Result as FmtResult};
 use std::fs::File;
 use std::path::Path;
-use std::str::FromStr;
 
 use regex::Regex;
 use serde::{Deserialize, Serialize};
@@ -13,212 +9,13 @@ use gerritbot_spark as spark;
 
 use super::BotError;
 
-#[derive(Debug, Clone)]
-struct Filter {
-    pub regex: Regex,
-    pub enabled: bool,
-}
+mod filter;
+mod flags;
+mod user;
 
-#[derive(Serialize, Deserialize)]
-struct FilterForSerialize<'a> {
-    regex: &'a str,
-    enabled: bool,
-}
-
-#[derive(Copy, Clone, Debug, Serialize, Deserialize, PartialEq, Eq, Hash)]
-#[serde(rename_all = "snake_case")]
-pub enum UserFlag {
-    /// User wants notification messages for reviews with approvals.
-    NotifyReviewApprovals,
-    /// User wants notification messages for review comments without approvals.
-    NotifyReviewComments,
-    /// User wants notification messages for reviews with inline comments.
-    NotifyReviewInlineComments,
-    /// User wants notification messages when added as reviewer to a change.
-    NotifyReviewerAdded,
-}
-
-impl Display for UserFlag {
-    fn fmt(&self, f: &mut Formatter) -> FmtResult {
-        if let Ok(serde_json::Value::String(s)) = serde_json::to_value(self) {
-            write!(f, "{}", s)
-        } else {
-            panic!("failed to encode flag")
-        }
-    }
-}
-
-impl FromStr for UserFlag {
-    type Err = serde_json::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        serde_json::from_slice(s.as_bytes())
-    }
-}
-
-/// Default flags for users that haven't enabled or disabled anything specific.
-const DEFAULT_FLAGS: &[UserFlag] = &[
-    UserFlag::NotifyReviewApprovals,
-    UserFlag::NotifyReviewInlineComments,
-    UserFlag::NotifyReviewerAdded,
-];
-
-/// All flags that deal with review comments.
-pub const REVIEW_COMMENT_FLAGS: &[UserFlag] = &[
-    UserFlag::NotifyReviewApprovals,
-    UserFlag::NotifyReviewComments,
-    UserFlag::NotifyReviewInlineComments,
-];
-
-/// All flags that deal with notifications.
-pub const NOTIFICATION_FLAGS: &[UserFlag] = &[
-    UserFlag::NotifyReviewApprovals,
-    UserFlag::NotifyReviewComments,
-    UserFlag::NotifyReviewInlineComments,
-    UserFlag::NotifyReviewerAdded,
-];
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-enum UserFlags {
-    Default,
-    // Note: this could be optimized into bitflags to make it faster and avoid
-    // allocation.
-    Custom(HashSet<UserFlag>),
-}
-
-impl Default for UserFlags {
-    fn default() -> Self {
-        UserFlags::Default
-    }
-}
-
-impl UserFlags {
-    fn contains(&self, flag: UserFlag) -> bool {
-        match self {
-            UserFlags::Default => DEFAULT_FLAGS
-                .iter()
-                .any(|default_flag| default_flag == &flag),
-            UserFlags::Custom(set) => set.contains(&flag),
-        }
-    }
-
-    fn is_default(&self) -> bool {
-        if let UserFlags::Default = self {
-            true
-        } else {
-            false
-        }
-    }
-
-    fn reset(&mut self) {
-        std::mem::replace(self, UserFlags::Default);
-    }
-
-    fn set(&mut self, flag: UserFlag, value: bool) {
-        let set_flag = |set: &mut HashSet<UserFlag>| {
-            if value {
-                set.insert(flag);
-            } else {
-                set.remove(&flag);
-            }
-        };
-
-        match self {
-            UserFlags::Default => {
-                let mut set = DEFAULT_FLAGS.iter().cloned().collect();
-                set_flag(&mut set);
-                std::mem::replace(self, UserFlags::Custom(set));
-            }
-            UserFlags::Custom(ref mut set) => set_flag(set),
-        }
-    }
-}
-
-/// Serialize the filter by storing the regex as a string.
-fn serialize_filter<S>(filter: &Option<Filter>, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    filter
-        .as_ref()
-        .map(|f| FilterForSerialize {
-            regex: f.regex.as_str(),
-            enabled: f.enabled,
-        })
-        .serialize(serializer)
-}
-
-/// Deserialize the filter by compiling the regex.
-fn deserialize_filter<'de, D>(deserializer: D) -> Result<Option<Filter>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let maybe_filter = Option::<FilterForSerialize>::deserialize(deserializer)?;
-
-    maybe_filter
-        .map(|f| {
-            Regex::new(f.regex)
-                .map(|regex| Filter {
-                    regex,
-                    enabled: f.enabled,
-                })
-                .map_err(|e| {
-                    <D::Error as serde::de::Error>::custom(format!("invalid regex: {}", e))
-                })
-        })
-        .transpose()
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct User {
-    // Legacy attribute.  Keep so we don't drop it on deserialize, serialize.
-    // Should be removed later.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    spark_person_id: Option<String>,
-    /// email of the user; assumed to be the same in Spark and Gerrit
-    email: spark::Email,
-    #[serde(skip_serializing_if = "UserFlags::is_default", default)]
-    flags: UserFlags,
-    enabled: bool,
-    #[serde(
-        skip_serializing_if = "Option::is_none",
-        serialize_with = "serialize_filter",
-        deserialize_with = "deserialize_filter",
-        default
-    )]
-    filter: Option<Filter>,
-}
-
-impl User {
-    fn new(email: spark::Email) -> Self {
-        Self {
-            spark_person_id: None,
-            email: email,
-            filter: None,
-            enabled: true,
-            flags: UserFlags::Default,
-        }
-    }
-
-    pub fn email(&self) -> &spark::EmailRef {
-        &self.email
-    }
-
-    pub fn has_any_flag<I, F>(&self, flags: I) -> bool
-    where
-        I: IntoIterator<Item = F>,
-        F: Borrow<UserFlag>,
-    {
-        self.enabled
-            && flags
-                .into_iter()
-                .any(|flag| self.flags.contains(*flag.borrow()))
-    }
-
-    pub fn has_flag(&self, flag: UserFlag) -> bool {
-        self.has_any_flag(&[flag])
-    }
-}
+use filter::Filter;
+pub use flags::{UserFlag, NOTIFICATION_FLAGS, REVIEW_COMMENT_FLAGS};
+pub use user::User;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct State {
